Fixing issue with script not able to parse description with whitespaces
If a user was specifying 'TEXT_BUILD_DESCRIPTION' in Jenkins with
whitespaces, the script was failing to pass this as a one string.
Instead, it was passing it as few strings.
To fix this:
- I put 'description' var in single quotes
- Trimming leading and trailing whitespaces in the declaration of the
'description' var. This is needed because the value of
currentBuild.description is already trimmed, so when comparing it to
 untrimmed 'description' we get two different values. This cause to the
same line to be written twice in the excel sheet. For example, if not
trimmed description, this val would be different on both vars:
" this is not good ". in the description var it would be as is, but in
currentBuild.description var it would be as "this is not good". It will
cause two duplicated lines to be created on the excel sheet.